Add the Django project structure
We're using django classy settings to split up environments. The
Development environment contains all the settings for development. The
production settings are configured in the API class.

Environment variables are configured in .env files. Production variables
are configured in the .env file. development variables are configured in
.env.development.

The requirements for the project are listed in requirements.txt while
the versions pinnings are set in constraints.txt